+++ src/test/java/org/kohsuke/github/GHProjectColumnTest.java
@@ -2,6 +2,7 @@ package org.kohsuke.github;
 
 import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.io.FileNotFoundException;
@@ -10,14 +11,14 @@ import java.io.IOException;
 /**
  * @author Gunnar Skjold
  */
-public class GHProjectColumnTest extends AbstractGitHubApiTestBase {
+public class GHProjectColumnTest extends AbstractGitHubApiWireMockTest {
 	private GHProject project;
 	private GHProjectColumn column;
 
-	@Override public void setUp() throws Exception {
-		super.setUp();
+	@Before
+	public void setUp() throws Exception {
 		project = gitHub
-				.getOrganization("github-api-test-org")
+				.getOrganization(GITHUB_API_TEST_ORG)
 				.createProject("test-project", "This is a test project");
 		column = project.createColumn("column-one");
 	}
@@ -47,20 +48,26 @@ public class GHProjectColumnTest extends AbstractGitHubApiTestBase {
 
 	@After
 	public void after() throws IOException {
-		if(column != null) {
-			try {
-				column.delete();
-				column = null;
-			} catch (FileNotFoundException e) {
-				column = null;
+		if(githubApi.isUseProxy()) {
+			if (column != null) {
+				column = gitHubBeforeAfter
+					.getProjectColumn(column.getId());
+				try {
+					column.delete();
+					column = null;
+				} catch (FileNotFoundException e) {
+					column = null;
+				}
 			}
-		}
-		if(project != null) {
-			try {
-				project.delete();
-				project = null;
-			} catch (FileNotFoundException e) {
-				project = null;
+			if (project != null) {
+				project = gitHubBeforeAfter
+					.getProject(project.getId());
+				try {
+					project.delete();
+					project = null;
+				} catch (FileNotFoundException e) {
+					project = null;
+				}
 			}
 		}
 	}
