+++ src/test/java/org/kohsuke/github/GHProjectTest.java
@@ -2,6 +2,7 @@ package org.kohsuke.github;
 
 import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.io.FileNotFoundException;
@@ -10,14 +11,14 @@ import java.io.IOException;
 /**
  * @author Gunnar Skjold
  */
-public class GHProjectTest extends AbstractGitHubApiTestBase {
+public class GHProjectTest extends AbstractGitHubApiWireMockTest {
 	private GHProject project;
 
-	@Override public void setUp() throws Exception {
-		super.setUp();
+	@Before
+	public void setUp() throws Exception {
 		project = gitHub
-				.getOrganization("github-api-test-org")
-				.createProject("test-project", "This is a test project");
+			.getOrganization(GITHUB_API_TEST_ORG)
+			.createProject("test-project", "This is a test project");
 	}
 
 	@Test
@@ -68,12 +69,16 @@ public class GHProjectTest extends AbstractGitHubApiTestBase {
 
 	@After
 	public void after() throws IOException {
-		if(project != null) {
-			try {
-				project.delete();
-				project = null;
-			} catch (FileNotFoundException e) {
-				project = null;
+		if (githubApi.isUseProxy()) {
+			if (project != null) {
+				project = gitHubBeforeAfter
+					.getProject(project.getId());
+				try {
+					project.delete();
+					project = null;
+				} catch (FileNotFoundException e) {
+					project = null;
+				}
 			}
 		}
 	}
